Migrate from setup.py to pyproject.toml
