+++ dev/docs/source/getting_started.rst
@@ -38,7 +38,7 @@ it as follows (change the version number to suit)::
     $ tar -zxvf XlsxWriter-1.2.3.tar.gz
 
     $ cd XlsxWriter-1.2.3
-    $ python setup.py install
+    $ pip install --user .
 
 A tarball of the latest code can be downloaded from GitHub as follows::
 
@@ -46,7 +46,7 @@ A tarball of the latest code can be downloaded from GitHub as follows::
 
     $ tar zxvf main.tar.gz
     $ cd XlsxWriter-main/
-    $ python setup.py install
+    $ pip install --user .
 
 
 Cloning from GitHub
@@ -59,7 +59,7 @@ You can clone the repository and install from it as follows::
     $ git clone https://github.com/jmcnamara/XlsxWriter.git
 
     $ cd XlsxWriter
-    $ python setup.py install
+    $ pip install --user .
 
 
 Running a sample program
