--- setup.cfg
@@ -1,3 +1,0 @@
-[wheel]
-universal = 0
-
